Realtime-dwls/requirements.txt
# pip install -r requirements.txt


Pillow == 9.1.0
PyMsgBox == 1.0.9
geocoder == 1.38.1
jsonschema == 4.4.0
opencv_python == 4.5.5.64
pafy == 0.5.5
paho_mqtt == 1.6.1
requests == 2.27.1
youtube-dl == 2020.12.2
pandas >= 1.1.4
PyYAML >= 5.3.1
tqdm >=4.41.0
matplotlib >= 3.2.2
seaborn >= 0.11.0







